fix(acp): expose authoritative session token totals / 修复 ACP 会话令牌重复统计
Problem
ACP session status exposed prompt and cache partition fields without an authoritative total. Hosts that summed every token field double-counted cached input, while reasoning tokens are also a completion subset.

Fix
Add totalTokens to the schema-v1 usage payload as promptTokens + completionTokens. Keep the cache and reasoning breakdown unchanged, and derive the value at wire time so restored legacy snapshots receive the same correct total.

Verification
- go test -count=1 ./internal/acp
- go vet ./internal/acp
- go run ./tools/repolint

Refs #8668
